pyproject.toml
[tool.poetry]
name = "openglossary"
version = "0.1.0"
description = ""
authors = ["falx <falx@falx.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
markdown = "^3.7"
loguru = "^0.7.2"
gunicorn = "^23.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
